requirements.txt
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
pandas>=2.2   # lo usaremos luego para la ingesta
pyarrow>=16.0 # idem, para futuros .parquet o acelerar IO
